Make TransientView JSON serializable
The format is a mix of JSONRPC-style class hinting and the JSON representation
of BSON objects. It can be deserialized back into a TransientView with
ampel.utils.json.load_objects(), but is descriptive enough to be used as-is by
3rd party code.

Some changes were necessary to make custom Ampel classes deserializable:
- Classes created by AmpelMetaFlags set their __module__ attribute
- Compound reimplemented as a dataclass and moved ampel-core to ampel-base

This introduces a dependency on the dataclasses module, added to the standard
library in Python 3.7. A backport exists for Python 3.6 and can be installed
with `pip`